refactor(poster): Create base poster module, and GEWIS+Hubble poster handlers
