+++ src/modules/handlers/screen/poster/base-poster-screen-handler.ts
@@ -4,7 +4,7 @@ import { TrackChangeEvent } from '../../../events/music-emitter-events';
 import { PosterManager } from './poster-manager';
 import { TrelloPosterManager } from './trello/trello-poster-manager';
 
-export default class PosterScreenHandler extends BaseScreenHandler {
+export default class BasePosterScreenHandler extends BaseScreenHandler {
   public posterManager: PosterManager;
 
   private borrelModeDay: number | undefined;
