+++ src/modules/handlers/screen/poster/gewis-poster-screen-controller.ts
@@ -1,0 +1,100 @@
+import {BasePosterResponse, BasePosterScreenController, BorrelModeParams} from "./base-poster-screen-controller";
+import HandlerManager from "../../../root/handler-manager";
+import {Screen} from "../../../root/entities";
+import BasePosterScreenHandler from "./base-poster-screen-handler";
+import {GewisPosterScreenHandler} from "../index";
+import {Body, Get, Post, Put, Query, Request, Route, Security, Tags} from "tsoa";
+import {SecurityNames} from "../../../../helpers/security";
+import {securityGroups} from "../../../../helpers/security-groups";
+import logger from "../../../../logger";
+import {Request as ExpressRequest} from "express";
+import NsTrainsService, {TrainResponse} from "./ns-trains-service";
+import GEWISPosterService, {GEWISPhotoAlbumParams} from "./gewis-poster-service";
+import OlympicsService from "./olympics-service";
+
+interface GewisPosterResponse extends BasePosterResponse {
+    borrelMode: boolean;
+}
+
+@Route('handler/screen/gewis-poster')
+@Tags('Handlers')
+export class GewisPosterScreenController extends BasePosterScreenController {
+    constructor() {
+        super();
+        this.screenHandler = HandlerManager.getInstance()
+            .getHandlers(Screen)
+            .filter((h) => h.constructor.name === GewisPosterScreenHandler.name)[0] as BasePosterScreenHandler;
+    }
+
+    @Security(SecurityNames.LOCAL, securityGroups.poster.base)
+    @Get('')
+    public async getGewisPosters(@Query() alwaysReturnBorrelPosters?: boolean): Promise<GewisPosterResponse> {
+        if (!this.screenHandler.posterManager.posters) {
+            try {
+                await this.screenHandler.posterManager.fetchPosters();
+            } catch (e) {
+                logger.error(e);
+            }
+        }
+        const posters = this.screenHandler.posterManager.posters ?? [];
+        if (alwaysReturnBorrelPosters || this.screenHandler.borrelMode) {
+            return {
+                posters,
+                borrelMode: this.screenHandler.borrelMode,
+            };
+        }
+        return {
+            posters: posters.filter((p) => !p.borrelMode),
+            borrelMode: false,
+        };
+    }
+
+    @Security(SecurityNames.LOCAL, securityGroups.poster.privileged)
+    @Post('force-update')
+    public async forceUpdateGewisPosters(@Request() req: ExpressRequest): Promise<void> {
+        super.forceUpdatePosters(req);
+    }
+
+    @Security(SecurityNames.LOCAL, securityGroups.poster.base)
+    @Get('borrel-mode')
+    public async getPosterBorrelMode(): Promise<BorrelModeParams> {
+        return { enabled: this.screenHandler.borrelMode };
+    }
+
+    @Security(SecurityNames.LOCAL, securityGroups.poster.base)
+    @Put('borrel-mode')
+    public async setPosterBorrelMode(
+        @Request() req: ExpressRequest,
+        @Body() body: BorrelModeParams,
+    ): Promise<void> {
+        logger.audit(
+            req.user,
+            `Set poster screen borrel mode to "${body.enabled ? 'true' : 'false'}".`,
+        );
+        this.screenHandler.setBorrelModeEnabled(body.enabled);
+    }
+
+    @Security(SecurityNames.LOCAL, securityGroups.poster.base)
+    @Get('train-departures')
+    public async getTrains(): Promise<TrainResponse[]> {
+        return new NsTrainsService().getTrains();
+    }
+
+    @Security(SecurityNames.LOCAL, securityGroups.poster.base)
+    @Post('photo')
+    public async getPhoto(@Body() params: GEWISPhotoAlbumParams) {
+        return new GEWISPosterService().getPhoto(params);
+    }
+
+    @Security(SecurityNames.LOCAL, securityGroups.poster.base)
+    @Get('olympics/medal-table')
+    public async getOlympicsMedalTable() {
+        return new OlympicsService().getMedalTable();
+    }
+
+    @Security(SecurityNames.LOCAL, securityGroups.poster.base)
+    @Get('olympics/country-medals')
+    public async getDutchOlympicMedals() {
+        return new OlympicsService().getDutchMedals();
+    }
+}
