+++ src/modules/handlers/screen/poster/gewis-poster-screen-handler.ts
@@ -1,0 +1,9 @@
+import BasePosterScreenHandler from "./base-poster-screen-handler";
+import { Namespace } from "socket.io";
+
+
+export default class GewisPosterScreenHandler extends BasePosterScreenHandler {
+    constructor(socket: Namespace) {
+        super(socket);
+    }
+}
