+++ src/modules/handlers/screen/poster/hubble-poster-screen-handler.ts
@@ -1,0 +1,8 @@
+import BasePosterScreenHandler from "./base-poster-screen-handler";
+import { Namespace } from "socket.io";
+
+export default class HubblePosterScreenHandler extends BasePosterScreenHandler {
+    constructor(socket: Namespace) {
+        super(socket);
+    }
+}
