+++ src/modules/handlers/screen/poster/index.ts
@@ -1,3 +1,3 @@
 // eslint-disable-next-line import/no-cycle -- TODO fix cyclic dependency
-export { PosterScreenController } from './poster-screen-controller';
-export { default as PosterScreenHandler } from './poster-screen-handler';
+export { BasePosterScreenController } from './base-poster-screen-controller';
+export { default as PosterScreenHandler } from './base-poster-screen-handler';
